+++ Example/DBSwiftTest/Database/Passenger/Passenger.swift
@@ -15,16 +15,16 @@ class Passenger: NSManagedObject {
 // Insert code here to add functionality to your managed object subclass
 
 	var allBaggagesCount : Int {
-		return self.mutableSetValueForKey(JSON.Baggages).count
+		return self.mutableSetValue(forKey: JSON.Baggages).count
 	}
 
 	var allBaggages : NSMutableSet {
-		return self.mutableSetValueForKey(JSON.Baggages)
+		return self.mutableSetValue(forKey: JSON.Baggages)
 	}
 
 	var allBaggagesArray : [Baggage] {
 		let sortDescriptor = NSSortDescriptor(key: JSON.Weight, ascending: true)
-		return (self.allBaggages.sortedArrayUsingDescriptors([sortDescriptor]) as? [Baggage] ?? [])
+		return (self.allBaggages.sortedArray(using: [sortDescriptor]) as? [Baggage] ?? [])
 	}
 }
 
@@ -33,7 +33,7 @@ class Passenger: NSManagedObject {
 extension Passenger {
 
 	override var description : String {
-		var description = "\(self.objectID.URIRepresentation().lastPathComponent ?? "")"
+		var description = "\(self.objectID.uriRepresentation().lastPathComponent)"
 		if let n = self.name {
 			description += ": \(n)"
 		}
@@ -42,7 +42,7 @@ extension Passenger {
 		}
 		description += ", \(self.allBaggagesCount) Baggage(s)"
 
-		if let plane = self.plane?.objectID.URIRepresentation().lastPathComponent {
+		if let plane = self.plane?.objectID.uriRepresentation().lastPathComponent {
 			description += ", plane: \(plane)"
 		}
 		return description
@@ -56,15 +56,15 @@ extension Passenger {
 		}
 	}
 
-	override func updateWithDictionary(dictionary: [NSObject : AnyObject]?, inContext savingContext: NSManagedObjectContext) {
-		super.updateWithDictionary(dictionary, inContext: savingContext)
+	override func update(with dictionary: [AnyHashable: Any]?, in savingContext: NSManagedObjectContext) {
+		super.update(with: dictionary, in: savingContext)
 
 		self.name = GET_STRING(dictionary, JSON.Name)
 		self.age = GET_NUMBER(dictionary, JSON.Age)
 
-		if let jsonArray = OBJECT(dictionary, JSON.Baggages) as? [[NSObject : AnyObject]] {
-			if let passengers = Baggage.crudEntitiesWithArray(jsonArray, inContext: savingContext) {
-				self.mutableSetValueForKey(JSON.Baggages).addObjectsFromArray(passengers)
+		if let jsonArray = OBJECT(dictionary, JSON.Baggages) as? [[AnyHashable: Any]] {
+			if let passengers = Baggage.crudEntities(with: jsonArray, in: savingContext) {
+				self.mutableSetValue(forKey: JSON.Baggages).addObjects(from: passengers)
 			}
 		}
 	}
@@ -75,11 +75,11 @@ extension Passenger {
 			return invalidReason
 		}
 
-		guard let name = self.name where (name.characters.count > 0) else {
+		guard let name = self.name, (name.characters.count > 0) else {
 			return "Invalid Name"
 		}
 
-		guard let age = self.age as? Int where (age > 0) else {
+		guard let age = self.age as? Int, (age > 0) else {
 			return "Invalid age. can't be inferior or equal to 0"
 		}
 
@@ -87,13 +87,13 @@ extension Passenger {
 		return nil
 	}
 
-	override func deleteEntityWithReason(reason: String?, inContext savingContext: NSManagedObjectContext) {
+	override func deleteEntity(withReason reason: String?, in savingContext: NSManagedObjectContext) {
 
 		for baggage in self.allBaggagesArray {
-			baggage.deleteEntityWithReason("Parent passenger entity removed", inContext: savingContext)
+			baggage.deleteEntity(withReason: "Parent passenger entity removed", in: savingContext)
 		}
 
-		super.deleteEntityWithReason(reason, inContext: savingContext)
+		super.deleteEntity(withReason: reason, in: savingContext)
 	}
 
 	override class func verbose() -> Bool {
@@ -104,16 +104,16 @@ extension Passenger {
 		return JSON.Name
 	}
 
-	override class func primaryPredicateWithDictionary(dictionary: [NSObject:AnyObject]?) -> NSPredicate? {
+	override class func primaryPredicate(with dictionary: [AnyHashable: Any]?) -> NSPredicate? {
 		if let
 			name = GET_STRING(dictionary, JSON.Name),
-			age = GET_NUMBER(dictionary, JSON.Age) {
+			let age = GET_NUMBER(dictionary, JSON.Age) {
 				return NSPredicate(format: "%K ==[c] %@ && %K ==[c] %@", JSON.Name, name, JSON.Age, age)
 		}
-		return super.primaryPredicateWithDictionary(dictionary)
+		return super.primaryPredicate(with: dictionary)
 	}
 
-	override func shouldUpdateEntityWithDictionary(dictionary: [NSObject : AnyObject]?, inContext savingContext: NSManagedObjectContext) -> Bool {
-		return super.shouldUpdateEntityWithDictionary(dictionary, inContext: savingContext)
+	override func shouldUpdateEntity(with dictionary: [AnyHashable: Any]?, in savingContext: NSManagedObjectContext) -> Bool {
+		return super.shouldUpdateEntity(with: dictionary, in: savingContext)
 	}
 }