+++ Example/DBSwiftTest/MockManager.swift
@@ -16,7 +16,7 @@ class MockManager {
 		var data = [[String:AnyObject]]()
 
 		for _ in 0...Int(arc4random_uniform(10) + 1) {
-			data.append([JSON.Origin: self.randomCity(), JSON.Destination: self.randomCity(), JSON.Passengers: self.randomPassengerJSON()])
+			data.append([JSON.Origin: self.randomCity() as AnyObject, JSON.Destination: self.randomCity() as AnyObject, JSON.Passengers: self.randomPassengerJSON() as AnyObject])
 		}
 
 		return data
@@ -26,7 +26,7 @@ class MockManager {
 		var data = [[String:AnyObject]]()
 
 		for _ in 0...Int(arc4random_uniform(10) + 1) {
-			data.append([JSON.Name: self.randomName(), JSON.Age: Int(arc4random_uniform(20)) + 20, JSON.Baggages: self.randomBaggageJSON()])
+			data.append([JSON.Name: self.randomName() as AnyObject, JSON.Age: Int(arc4random_uniform(20)) + 20 as AnyObject, JSON.Baggages: self.randomBaggageJSON() as AnyObject])
 		}
 
 		return data
@@ -36,7 +36,7 @@ class MockManager {
 		var data = [[String:AnyObject]]()
 
 		for _ in 0...Int(arc4random_uniform(2) + 1) {
-			data.append([JSON.Weight: Int(arc4random_uniform(20)) - 2])
+			data.append([JSON.Weight: Int(arc4random_uniform(20)) - 2 as AnyObject])
 		}
 
 		return data
@@ -44,21 +44,21 @@ class MockManager {
 
 	// MARK: - Private Methods
 
-	private class func randomCity() -> String {
+	fileprivate class func randomCity() -> String {
 		let cities = self.cities()
 		return cities[Int(arc4random_uniform(UInt32(cities.count)))]
 	}
 
-	private class func cities() -> [String] {
+	fileprivate class func cities() -> [String] {
 		return ["Paris", "London", "Berlin", "Tokyo", "Madrid", "Los Angles", "Toronto", "Sydney", "Hong Kong"]
 	}
 
-	private class func randomName() -> String {
+	fileprivate class func randomName() -> String {
 		let names = self.names()
 		return names[Int(arc4random_uniform(UInt32(names.count)))]
 	}
 
-	private class func names() -> [String] {
+	fileprivate class func names() -> [String] {
 		return ["Kevin", "Michael", "James", "John", "Bob", "Pierre", "Alex", "Tom", "Jack", "Nicolas"]
 	}
 }