+++ Example/DKDBManagerTests/DeleteFunctionTests.swift
@@ -25,13 +25,13 @@ extension DeleteFunctionTest {
 	func testShouldDeleteAllEntitiesInContext() {
 
 		// set expectation
-		let expectation = self.expectationWithDescription("Wait for the Response")
+		let expectation = self.expectation(description: "Wait for the Response")
 
-		TestDataManager.saveWithBlock({ (savingContext: NSManagedObjectContext) -> Void in
+		TestDataManager.save({ (savingContext: NSManagedObjectContext) -> Void in
 			// background thread
-			TestDataManager.deleteAllEntitiesInContext(savingContext)
+			TestDataManager.deleteAllEntities(in: savingContext)
 
-		}) { (contextDidSave: Bool, error: NSError?) -> Void in
+		}, completion: { (contextDidSave: Bool, error: Error?) -> Void in
 			// main thread
 			XCTAssertTrue(contextDidSave)
 			XCTAssertNil(error)
@@ -46,25 +46,24 @@ extension DeleteFunctionTest {
 			XCTAssertEqual(Baggage.all()?.count, 0, "the number of baggages should be equals to 0")
 
 			expectation.fulfill()
-		}
+		})
 
-		self.waitForExpectationsWithTimeout(5, handler: nil)
-
+		self.waitForExpectations(timeout: 5, handler: nil)
 	}
 
 	func testShouldDeleteAllPlanesAndSubEntities() {
 
 		// set expectation
-		let expectation = self.expectationWithDescription("Wait for the Response")
+		let expectation = self.expectation(description: "Wait for the Response")
 
-		TestDataManager.saveWithBlock({ (savingContext: NSManagedObjectContext) -> Void in
+		TestDataManager.save({ (savingContext: NSManagedObjectContext) -> Void in
 			// background thread
-			let planes = (Plane.MR_findAllInContext(savingContext) as? [Plane] ?? [])
+			let planes = (Plane.mr_findAll(in: savingContext) as? [Plane] ?? [])
 			for plane in planes {
-				plane.deleteEntityWithReason(nil, inContext: savingContext)
+				plane.deleteEntity(withReason: nil, in: savingContext)
 			}
 
-		}) { (contextDidSave: Bool, error: NSError?) -> Void in
+		}, completion: { (contextDidSave: Bool, error: Error?) -> Void in
 			// main thread
 			XCTAssertTrue(contextDidSave)
 			XCTAssertNil(error)
@@ -79,24 +78,24 @@ extension DeleteFunctionTest {
 			XCTAssertEqual(Baggage.all()?.count, 0, "the number of baggages should be equals to 0")
 
 			expectation.fulfill()
-		}
+		})
 
-		self.waitForExpectationsWithTimeout(5, handler: nil)
+		self.waitForExpectations(timeout: 5, handler: nil)
 	}
 
 	func testShouldDeleteAllBaggagesButNotParentEntities() {
 
 		// set expectation
-		let expectation = self.expectationWithDescription("Wait for the Response")
+		let expectation = self.expectation(description: "Wait for the Response")
 
-		TestDataManager.saveWithBlock({ (savingContext: NSManagedObjectContext) -> Void in
+		TestDataManager.save({ (savingContext: NSManagedObjectContext) -> Void in
 			// background thread
-			let baggages = (Baggage.MR_findAllInContext(savingContext) as? [Baggage] ?? [])
+			let baggages = (Baggage.mr_findAll(in: savingContext) as? [Baggage] ?? [])
 			for baggage in baggages {
-				baggage.deleteEntityWithReason(nil, inContext: savingContext)
+				baggage.deleteEntity(withReason: nil, in: savingContext)
 			}
 
-		}) { (contextDidSave: Bool, error: NSError?) -> Void in
+		}, completion: { (contextDidSave: Bool, error: Error?) -> Void in
 			// main thread
 			XCTAssertTrue(contextDidSave)
 			XCTAssertNil(error)
@@ -111,21 +110,21 @@ extension DeleteFunctionTest {
 			XCTAssertEqual(Baggage.all()?.count, 0, "the number of baggages should be equals to 0")
 
 			expectation.fulfill()
-		}
+		})
 
-		self.waitForExpectationsWithTimeout(5, handler: nil)
+		self.waitForExpectations(timeout: 5, handler: nil)
 	}
 
 	func testShouldDeleteEntitiesForClassAndNotSubEntities() {
 
 		// set expectation
-		let expectation = self.expectationWithDescription("Wait for the Response")
+		let expectation = self.expectation(description: "Wait for the Response")
 
-		TestDataManager.saveWithBlock({ (savingContext: NSManagedObjectContext) -> Void in
+		TestDataManager.save({ (savingContext: NSManagedObjectContext) -> Void in
 			// background thread
-			TestDataManager.deleteAllEntitiesForClass(Plane.self, inContext: savingContext)
+			TestDataManager.deleteAllEntities(for: Plane.self, in: savingContext)
 
-		}) { (contextDidSave: Bool, error: NSError?) -> Void in
+		}, completion: { (contextDidSave: Bool, error: Error?) -> Void in
 			// main thread
 			XCTAssertTrue(contextDidSave)
 			XCTAssertNil(error)
@@ -140,9 +139,9 @@ extension DeleteFunctionTest {
 			XCTAssertNotEqual(Baggage.all()?.count, 0, "the number of baggages should NOT be equals to 0")
 
 			expectation.fulfill()
-		}
+		})
 
-		self.waitForExpectationsWithTimeout(5, handler: nil)
+		self.waitForExpectations(timeout: 5, handler: nil)
 	}
 }